% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deciling function.R
\name{deciling}
\alias{deciling}
\title{SIC Profiling}
\usage{
deciling(access_token, instance_url, object, catfield, numfield, non_core,
  dynamic, data_split, newname)
}
\description{
This function indexes one dependent variable from a numeric variable
}
