% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Focus.R
\name{focus}
\alias{focus}
\title{Industry ranking}
\usage{
focus(access_token, instance_url, object, depfield, indfield, newname)
}
\description{
This function ranks a categorical variable from a numeric variable
}
