% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Future_recency.R
\name{Future_recency}
\alias{Future_recency}
\title{Past Recency}
\usage{
Future_recency(access_token, instance_url, object, field1, newname)
}
\description{
This function will do the past recency of the date variable
}
