% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/past_date_recency_new.R
\name{past_date_recency_new}
\alias{past_date_recency_new}
\title{Past Recency}
\usage{
past_date_recency_new(data5, b)
}
\description{
This function will do the past recency of the date variable
}
