% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/past_date_recency_old.R
\name{past_date_recency_old}
\alias{past_date_recency_old}
\title{Past Recency}
\usage{
past_date_recency_old(data5, b)
}
\description{
This function will do the past recency of the date variable
}
