% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/past_recency.R
\name{past_recency}
\alias{past_recency}
\title{Past Recency}
\usage{
past_recency(access_token, instance_url, object, field1, newname)
}
\description{
This function will do the past recency of the date variable
}
