% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wisdom.R
\name{predictor}
\alias{predictor}
\title{Predictive Profiling}
\usage{
predictor(access_token, instance_url, object, numfield, catfield, newname)
}
\description{
This function indexes one dependent variable from a numeric variable
}
