% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocity.R
\name{quad}
\alias{quad}
\title{This function takes two variables and converts them into a quad}
\usage{
quad(access_token, instance_url, object, field1, field2, newname)
}
\description{
This function takes two variables and converts them into a quad
}
