% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quad_new.R
\name{quad_new}
\alias{quad_new}
\title{This function takes two variables and converts them into a quad}
\usage{
quad_new(access_token, instance_url, object, field1, data_type1, field2,
  data_type2, newname)
}
\description{
This function takes two variables and converts them into a quad
}
