% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Focus.R, R/ranker.R
\name{ranks}
\alias{ranks}
\alias{ranks}
\title{Industry ranking}
\usage{
ranks(access_token, instance_url, object, depfield, indfield, newname)

ranks(access_token, instance_url, object, depfield, indfield, newname)
}
\description{
This function ranks a categorical variable from a numeric variable

This function ranks a categorical variable from a numeric variable
}
