% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_status.R
\name{segment_new}
\alias{segment_new}
\title{Segment}
\usage{
segment_new(access_token, instance_url, object, field, newname, data_type)
}
\description{
This function builds a categorical variable from a numeric variable
}
