% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_status.R
\name{segment_old}
\alias{segment_old}
\title{Segment}
\usage{
segment_old(access_token, instance_url, object, field, data_type, newname)
}
\description{
This function builds a categorical variable from a numeric variable
}
