% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status_new.R
\name{status_new}
\alias{status_new}
\title{Segment}
\usage{
status_new(access_token, instance_url, object, field, newname)
}
\description{
This function builds a categorical variable from a numeric variable
}
