% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_wisdom.R
\name{wisdom}
\alias{wisdom}
\title{Predictive Profiling}
\usage{
wisdom(access_token, instance_url, object, numfield, data_type1, catfield,
  data_type2, newname)
}
\description{
This function indexes one dependent variable from a numeric variable
}
